Moved NavBar to new ui folder
